backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.0
